requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
six==1.17.0
supabase==2.10.0
typing_extensions==4.15.0
urllib3==2.5.0
yarl==1.22.0